Potential fix for code scanning alert no. 4

+++ core/arch/arm/kernel/secure_partition.c
@@ -1021,7 +1021,7 @@ static TEE_Result handle_fdt_dev_regions(struct sp_ctx *ctx, void *fdt)
 			return res;
 		}
 
-		res = vm_map(&ctx->uctx, &va, pages_cnt * SMALL_PAGE_SIZE,
+		res = vm_map(&ctx->uctx, &va, (size_t)pages_cnt * SMALL_PAGE_SIZE,
 			     perm, 0, m, 0);
 		mobj_put(m);
 		if (res)
